+++ undo/src/stream/mod.rs
@@ -1,6 +1,7 @@
 mod init;
 mod process;
 mod topics;
+mod undo;
 mod xack;
 mod xadd;
 mod xread;
